closes issue #10
Now the app uses BigDecimal to represent decimal values, and stores the as a String

+++ app/src/main/java/com/daniel/bugdetapp/Transaction.java
@@ -6,6 +6,8 @@ import androidx.room.Entity;
 import androidx.room.PrimaryKey;
 import org.threeten.bp.*;
 
+import java.math.BigDecimal;
+
 @Entity(tableName = "TRANSACTIONS")
 public class Transaction {
 
@@ -16,13 +18,13 @@ public class Transaction {
 
     @NonNull
     @ColumnInfo(name = "quantity")
-    private float quantity;
+    private BigDecimal quantity;
 
     @NonNull
     @ColumnInfo(name = "timestamp")
     private String timestamp;
 
-    public Transaction(@NonNull float quantity) {
+    public Transaction(@NonNull BigDecimal quantity) {
         this.quantity = quantity;
         this.timestamp = LocalDate.now().toString();
     }
@@ -31,7 +33,7 @@ public class Transaction {
         return key;
     }
 
-    public float getQuantity() {
+    public BigDecimal getQuantity() {
         return quantity;
     }
